move processing endpoint to email_routes.py bc of api_call_finished global variable
